src/search_head/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
english-words==1.0.3
filelock==3.0.12
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kafka-python==2.0.2
MarkupSafe==1.1.1
nltk==3.5
numpy==1.19.2
pymongo==3.11.0
PyYAML==5.3.1
regex==2020.10.23
requests==2.24.0
requests-file==1.5.1
six==1.15.0
soupsieve==2.0.1
tldextract==3.0.2
tqdm==4.51.0
urllib3==1.25.11
Werkzeug==1.0.1
